+++ examples/cis3-nft-sponsored-txs/src/lib.rs
@@ -39,21 +39,21 @@
 //! signed message.
 //!
 //! Concordium supports natively multi-sig accounts. Each account address on
-//! Concordium is controlled by one (or several) credentials (real-world
-//! identities) and each credential has one (or several) public-private key
-//! pairs. The key/signature structures of an account are represented as
-//! two-level maps (e.g. BTreeMap<CredentialIndexU8, BTreeMap<PublicKeyIndexU8,
-//! PublicKey>> or BTreeMap<CredentialIndexU8, BTreeMap<SignatureIndexU8,
-//! Signature>>). The outer map of `BTreeMap<CredentialIndexU8,
-//! BTreeMap<PublicKeyIndexU8, PublicKey>>` has an `AccountThreshold` (number of
-//! credentials needed to sign the transaction initiated by that account) and
-//! the inner map has a `SignatureThreshold` (number of Signatures needed for a
-//! specific credential so that this credential is considered to have signed the
-//! transaction initiated by that account). The CIS3 standard supports multi-sig
-//! accounts. But for simplicity, this contract supports only basic accounts
-//! that have exactly one credential and exactly one public key for that
-//! credential. As a result, all signaturesMaps/publicKeyMaps in this contract,
-//! will have only one value at key 0 in the inner and outer maps.
+//! Concordium is controlled by one or several credential(s) (real-world
+//! identities) and each credential has one or several public-private key
+//! pair(s). Both the key and signature structures of an account are represented
+//! as two-level maps (e.g. BTreeMap<CredentialIndexU8,
+//! BTreeMap<PublicKeyIndexU8, PublicKey>> or BTreeMap<CredentialIndexU8,
+//! BTreeMap<SignatureIndexU8, Signature>>). The outer map of
+//! `BTreeMap<CredentialIndexU8, BTreeMap<PublicKeyIndexU8, PublicKey>>` has an
+//! `AccountThreshold` (number of credentials needed to sign the transaction
+//! initiated by that account) and the inner map has a `SignatureThreshold`
+//! (number of Signatures needed for a specific credential so that this
+//! credential is considered to have signed the transaction initiated by that
+//! account). The CIS3 standard supports multi-sig accounts. For basic accounts
+//! (that have exactly one credential and exactly one public key for that
+//! credential), the signaturesMaps/publicKeyMaps in this contract, will have
+//! only one value at key 0 in the inner and outer maps.
 #![cfg_attr(not(feature = "std"), no_std)]
 
 use concordium_cis2::*;
@@ -275,11 +275,7 @@ pub struct PermitMessage {
 /// Takes a signature, the signer, and the message that was signed.
 #[derive(Serialize, SchemaType)]
 pub struct PermitParam {
-    /// Signature/s. The CIS3 standard supports multi-sig accounts. But for
-    /// simplicity, this contract only supports signatures that were generated
-    /// by accounts with exactly one public key (one credential with one public
-    /// key for that credential). The signature has to be at the key 0 in
-    /// both maps of the two-level map `AccountSignatures`.
+    /// Signature/s. The CIS3 standard supports multi-sig accounts.
     pub signature: AccountSignatures,
     /// Account that created the above signature.
     pub signer:    AccountAddress,
@@ -289,11 +285,7 @@ pub struct PermitParam {
 
 #[derive(Serialize)]
 pub struct PermitParamPartial {
-    /// Signature/s. The CIS3 standard supports multi-sig accounts. But for
-    /// simplicity, this contract only supports signatures that were generated
-    /// by accounts with exactly one public key (one credential with one public
-    /// key for that credential). The signature has to be at the key 0 in
-    /// both maps of the two-level map `AccountSignatures`.
+    /// Signature/s. The CIS3 standard supports multi-sig accounts.
     signature: AccountSignatures,
     /// Account that created the above signature.
     signer:    AccountAddress,
@@ -336,8 +328,6 @@ enum CustomContractError {
     WrongEntryPoint,
     /// Failed signature verification: Signature is expired.
     Expired,
-    /// Failed signature verification: Signature map is misconfigured.
-    SignatureMapMisconfigured,
 }
 
 /// Wrapping the custom errors in a type with CIS2 errors.
@@ -561,10 +551,11 @@ fn contract_view<S: HasStateApi>(
 
     let all_tokens = state.all_tokens.iter().map(|x| *x).collect();
 
-    let mut all_nonces = Vec::new();
-    for (account_address, nonce) in state.nonces_registry.iter() {
-        all_nonces.push((*account_address, *nonce));
-    }
+    let all_nonces: Vec<_> = state
+        .nonces_registry
+        .iter()
+        .map(|(account_address, nonce)| (*account_address, *nonce))
+        .collect();
 
     Ok(ViewState {
         state: inner_state,
@@ -796,7 +787,7 @@ fn contract_view_message_hash<S: HasStateApi>(
 ///   if:
 ///     - The `token_id` does not exist.
 ///     - The token is not owned by the `from` address.
-///     - The receive hook function calls rejects.
+///     - The receive hook function call rejects.
 #[receive(
     contract = "cis3_nft",
     name = "permit",
@@ -813,19 +804,6 @@ fn contract_permit<S: HasStateApi>(
 ) -> ContractResult<()> {
     // Parse the parameter.
     let param: PermitParam = ctx.parameter_cursor().get()?;
-
-    ensure!(
-        param.signature.sigs.len() == 1
-            && param
-                .signature
-                .sigs
-                .get(&0)
-                .ok_or(CustomContractError::SignatureMapMisconfigured)?
-                .sigs
-                .len()
-                == 1,
-        CustomContractError::SignatureMapMisconfigured.into()
-    );
 
     // Update the nonce.
     let mut entry = host.state_mut().nonces_registry.entry(param.signer).or_insert_with(|| 0);
@@ -1028,6 +1006,7 @@ fn contract_balance_of<S: HasStateApi>(
 
 /// Response type for the function `publicKeyOf`.
 #[derive(Debug, Serialize, SchemaType)]
+#[concordium(transparent)]
 pub struct PublicKeyOfQueryResponse(
     #[concordium(size_length = 2)] pub Vec<Option<AccountPublicKeys>>,
 );
@@ -1041,17 +1020,11 @@ impl From<Vec<Option<AccountPublicKeys>>> for PublicKeyOfQueryResponse {
 /// The parameter type for the contract functions `publicKeyOf/noneOf`. A query
 /// for the public key/nonce of a given account.
 #[derive(Debug, Serialize, SchemaType)]
+#[concordium(transparent)]
 pub struct VecOfAccountAddresses {
     /// List of queries.
     #[concordium(size_length = 2)]
-    pub queries: Vec<AccountAddressStruct>,
-}
-
-/// Part of the parameter type for the contract function `publicKeyOf/nonceOf`.
-#[derive(Debug, Serialize, SchemaType)]
-pub struct AccountAddressStruct {
-    /// The account for which the public key/nonce should be queried.
-    pub account: AccountAddress,
+    pub queries: Vec<AccountAddress>,
 }
 
 /// Get the public keys of accounts. `None` is returned if the account does not
@@ -1074,9 +1047,9 @@ fn contract_public_key_of<S: HasStateApi>(
     let params: VecOfAccountAddresses = ctx.parameter_cursor().get()?;
     // Build the response.
     let mut response: Vec<Option<AccountPublicKeys>> = Vec::with_capacity(params.queries.len());
-    for query in params.queries {
+    for account in params.queries {
         // Query the public_key.
-        let public_keys = host.account_public_keys(query.account).ok();
+        let public_keys = host.account_public_keys(account).ok();
 
         response.push(public_keys);
     }
@@ -1086,6 +1059,7 @@ fn contract_public_key_of<S: HasStateApi>(
 
 /// Response type for the function `nonceOf`.
 #[derive(Debug, Serialize, SchemaType)]
+#[concordium(transparent)]
 pub struct NonceOfQueryResponse(#[concordium(size_length = 2)] pub Vec<u64>);
 
 impl From<Vec<u64>> for NonceOfQueryResponse {
@@ -1111,10 +1085,9 @@ fn contract_nonce_of<S: HasStateApi>(
     let params: VecOfAccountAddresses = ctx.parameter_cursor().get()?;
     // Build the response.
     let mut response: Vec<u64> = Vec::with_capacity(params.queries.len());
-    for query in params.queries {
+    for account in params.queries {
         // Query the next nonce.
-        let nonce =
-            host.state().nonces_registry.get(&query.account).map(|nonce| *nonce).unwrap_or(0);
+        let nonce = host.state().nonces_registry.get(&account).map(|nonce| *nonce).unwrap_or(0);
 
         response.push(nonce);
     }